+++ changelog.md
@@ -5,6 +5,7 @@ Latest version: 0.31.2
 # Not released
 
 * [Added] Ability to remove some colors from the image by Color Filter
+* [Fixed] Empty manual lobule segmentation for central vein if no central is present
 
 # 0.31